+++ tilings/strategies/__init__.py
@@ -21,6 +21,7 @@ from .rearrange_assumption import RearrangeAssumptionFactory
 from .requirement_insertion import (
     CellInsertionFactory,
     FactorInsertionFactory,
+    PositiveCorroborationFactory,
     RemoveRequirementFactory,
     RequirementCorroborationFactory,
     RequirementExtensionFactory,
@@ -56,9 +57,10 @@ __all__ = [
     "RearrangeAssumptionFactory",
     "SplittingStrategy",
     # Batch
+    "AllPlacementsFactory",
     "CellInsertionFactory",
     "FactorInsertionFactory",
-    "AllPlacementsFactory",
+    "PositiveCorroborationFactory",
     "RemoveRequirementFactory",
     "RequirementExtensionFactory",
     "RequirementInsertionFactory",